--- a/pkg/streamingpromql/operators/vector_vector_binary_operation.go
+++ b/pkg/streamingpromql/operators/vector_vector_binary_operation.go
@@ -9,7 +9,6 @@ import (
 	"context"
 	"fmt"
 	"math"
-	"slices"
 	"sort"
 	"time"
 
@@ -394,37 +393,6 @@ func (b *VectorVectorBinaryOperation) groupLabelsFunc() func(labels.Labels) labe
 		lb.Del(labels.MetricName)
 		lb.Del(b.VectorMatching.MatchingLabels...)
 		return lb.Labels()
-	}
-}
-
-// vectorMatchingGroupKeyFunc returns a function that computes the grouping key of the output group a series belongs to.
-//
-// The return value from the function is valid until it is called again.
-func vectorMatchingGroupKeyFunc(vectorMatching parser.VectorMatching) func(labels.Labels) []byte {
-	buf := make([]byte, 0, 1024)
-
-	if vectorMatching.On {
-		slices.Sort(vectorMatching.MatchingLabels)
-
-		return func(l labels.Labels) []byte {
-			return l.BytesWithLabels(buf, vectorMatching.MatchingLabels...)
-		}
-	}
-
-	if len(vectorMatching.MatchingLabels) == 0 {
-		// Fast path for common case for expressions like "a + b" with no 'on' or 'without' labels.
-		return func(l labels.Labels) []byte {
-			return l.BytesWithoutLabels(buf, labels.MetricName)
-		}
-	}
-
-	lbls := make([]string, 0, len(vectorMatching.MatchingLabels)+1)
-	lbls = append(lbls, labels.MetricName)
-	lbls = append(lbls, vectorMatching.MatchingLabels...)
-	slices.Sort(lbls)
-
-	return func(l labels.Labels) []byte {
-		return l.BytesWithoutLabels(buf, lbls...)
 	}
 }