--- a/tools/find-unpooled-slice-creation.sh
+++ b/tools/find-unpooled-slice-creation.sh
@@ -1,4 +1,5 @@
 #! /usr/bin/env bash
+# SPDX-License-Identifier: AGPL-3.0-only
 
 set -euo pipefail