… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/streamingpromql/benchmarks/benchmarks.go
… pkg/streamingpromql/engine_concurrency_test.go
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/query.go
… pkg/streamingpromql/testdata/ours/binary_operators.test
… pkg/streamingpromql/testdata/upstream/operators.test