#### What this PR does

This PR adds support for the `and` binary operation.

Compared to Prometheus' engine, latency is up to 90% lower and peak memory consumption is up to 48% lower in our benchmarks:

```
                                                                       │  Prometheus   │                Mimir                │
                                                                       │    sec/op     │    sec/op     vs base               │
Query/a_1_and_b_1{l=~'.*[0-4]$'},_instant_query-10                        723.5µ ±  0%   717.8µ ±  1%        ~ (p=0.310 n=6)
Query/a_1_and_b_1{l=~'.*[0-4]$'},_range_query_with_100_steps-10           744.2µ ±  2%   734.0µ ±  4%        ~ (p=0.818 n=6)
Query/a_1_and_b_1{l=~'.*[0-4]$'},_range_query_with_1000_steps-10          1.269m ±  1%   1.130m ±  1%  -10.93% (p=0.002 n=6)
Query/a_100_and_b_100{l=~'.*[0-4]$'},_instant_query-10                    1.613m ± 11%   1.609m ±  2%        ~ (p=0.485 n=6)
Query/a_100_and_b_100{l=~'.*[0-4]$'},_range_query_with_100_steps-10       3.270m ±  2%   2.367m ± 10%  -27.62% (p=0.002 n=6)
Query/a_100_and_b_100{l=~'.*[0-4]$'},_range_query_with_1000_steps-10     17.321m ±  1%   8.414m ±  0%  -51.42% (p=0.002 n=6)
Query/a_2000_and_b_2000{l=~'.*[0-4]$'},_instant_query-10                  17.08m ±  1%   16.97m ±  2%        ~ (p=0.589 n=6)
Query/a_2000_and_b_2000{l=~'.*[0-4]$'},_range_query_with_100_steps-10     51.59m ±  4%   30.85m ±  1%  -40.20% (p=0.002 n=6)
Query/a_2000_and_b_2000{l=~'.*[0-4]$'},_range_query_with_1000_steps-10    392.2m ±  1%   146.9m ±  9%  -62.55% (p=0.002 n=6)
Query/a_1_and_b_1{l='notfound'},_instant_query-10                         232.3µ ±  2%   217.7µ ± 12%        ~ (p=0.065 n=6)
Query/a_1_and_b_1{l='notfound'},_range_query_with_100_steps-10            251.5µ ±  7%   218.5µ ±  6%  -13.12% (p=0.002 n=6)
Query/a_1_and_b_1{l='notfound'},_range_query_with_1000_steps-10           382.2µ ±  3%   235.5µ ±  1%  -38.38% (p=0.002 n=6)
Query/a_100_and_b_100{l='notfound'},_instant_query-10                     878.4µ ±  4%   524.2µ ±  5%  -40.32% (p=0.002 n=6)
Query/a_100_and_b_100{l='notfound'},_range_query_with_100_steps-10       1429.2µ ±  0%   570.8µ ±  1%  -60.06% (p=0.002 n=6)
Query/a_100_and_b_100{l='notfound'},_range_query_with_1000_steps-10       6.188m ±  3%   1.039m ±  2%  -83.22% (p=0.002 n=6)
Query/a_2000_and_b_2000{l='notfound'},_instant_query-10                  10.842m ±  1%   5.211m ±  1%  -51.93% (p=0.002 n=6)
Query/a_2000_and_b_2000{l='notfound'},_range_query_with_100_steps-10     21.741m ±  1%   5.582m ±  4%  -74.33% (p=0.002 n=6)
Query/a_2000_and_b_2000{l='notfound'},_range_query_with_1000_steps-10    144.98m ±  2%   12.39m ±  2%  -91.45% (p=0.002 n=6)
geomean                                                                   4.140m         2.229m        -46.17%

                                                                       │  Prometheus   │                Mimir                │
                                                                       │       B       │      B        vs base               │
Query/a_1_and_b_1{l=~'.*[0-4]$'},_instant_query-10                        70.23Mi ± 2%   69.00Mi ± 3%        ~ (p=0.143 n=6)
Query/a_1_and_b_1{l=~'.*[0-4]$'},_range_query_with_100_steps-10           69.90Mi ± 1%   70.33Mi ± 3%        ~ (p=0.589 n=6)
Query/a_1_and_b_1{l=~'.*[0-4]$'},_range_query_with_1000_steps-10          67.97Mi ± 1%   66.34Mi ± 4%        ~ (p=0.087 n=6)
Query/a_100_and_b_100{l=~'.*[0-4]$'},_instant_query-10                    66.05Mi ± 2%   66.66Mi ± 1%        ~ (p=0.420 n=6)
Query/a_100_and_b_100{l=~'.*[0-4]$'},_range_query_with_100_steps-10       66.83Mi ± 1%   66.73Mi ± 1%        ~ (p=0.394 n=6)
Query/a_100_and_b_100{l=~'.*[0-4]$'},_range_query_with_1000_steps-10      71.63Mi ± 1%   69.52Mi ± 2%   -2.96% (p=0.002 n=6)
Query/a_2000_and_b_2000{l=~'.*[0-4]$'},_instant_query-10                  68.12Mi ± 2%   69.08Mi ± 2%        ~ (p=0.310 n=6)
Query/a_2000_and_b_2000{l=~'.*[0-4]$'},_range_query_with_100_steps-10     83.23Mi ± 1%   72.56Mi ± 1%  -12.82% (p=0.002 n=6)
Query/a_2000_and_b_2000{l=~'.*[0-4]$'},_range_query_with_1000_steps-10    198.1Mi ± 1%   105.3Mi ± 4%  -46.83% (p=0.002 n=6)
Query/a_1_and_b_1{l='notfound'},_instant_query-10                         73.99Mi ± 2%   74.73Mi ± 1%        ~ (p=0.180 n=6)
Query/a_1_and_b_1{l='notfound'},_range_query_with_100_steps-10            73.16Mi ± 1%   74.46Mi ± 1%   +1.78% (p=0.004 n=6)
Query/a_1_and_b_1{l='notfound'},_range_query_with_1000_steps-10           70.31Mi ± 1%   73.52Mi ± 1%   +4.57% (p=0.002 n=6)
Query/a_100_and_b_100{l='notfound'},_instant_query-10                     67.06Mi ± 1%   67.49Mi ± 1%        ~ (p=0.093 n=6)
Query/a_100_and_b_100{l='notfound'},_range_query_with_100_steps-10        67.25Mi ± 3%   67.76Mi ± 1%        ~ (p=0.240 n=6)
Query/a_100_and_b_100{l='notfound'},_range_query_with_1000_steps-10       69.25Mi ± 1%   67.46Mi ± 1%   -2.58% (p=0.002 n=6)
Query/a_2000_and_b_2000{l='notfound'},_instant_query-10                   67.02Mi ± 2%   67.88Mi ± 2%        ~ (p=0.180 n=6)
Query/a_2000_and_b_2000{l='notfound'},_range_query_with_100_steps-10      75.27Mi ± 1%   68.97Mi ± 1%   -8.37% (p=0.002 n=6)
Query/a_2000_and_b_2000{l='notfound'},_range_query_with_1000_steps-10    129.24Mi ± 3%   67.18Mi ± 6%  -48.02% (p=0.002 n=6)
geomean                                                                   77.06Mi        70.97Mi        -7.90%
```

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [n/a] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.