@charleskorn 2024-10-04
> Thanks. I wonder if those re-uses have improved the benchmarks even more :-)

Not dramatically - reusing the left series slice for data [improved things by 1-2%](https://github.com/grafana/mimir/pull/9507#discussion_r1787013480), and reusing the metadata slice made no difference (within the margin of error of the benchmarks).